refactor: moves toc listener to own function
