+++ utils/index.ts
@@ -4,3 +4,4 @@ export * from './readFile'
 export * from './getDynamicPath'
 export * from './slugify'
 export * from './urls'
+export * from './toc_helpers'
